@ refactor(mcp): extract sql_modes and graph_data_mapping modules
- Phase 2: Move _handle_sql_mode from server.py to new tools/sql_modes.py
- Phase 3: Move graph_data_to_change_plan and 3 internal helpers from
  manage_graph_data.py to new tools/graph_data_mapping.py
- manage_graph_data.py re-exports graph_data_to_change_plan for backwards
  compatibility
- No behavior changes, all 226 tests pass

@
